rename ai crate to assistant crate
