+++ crates/assistant/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "ai"
+name = "assistant"
 version = "0.1.0"
 edition = "2021"
 publish = false
 
 [lib]
-path = "src/ai.rs"
+path = "src/assistant.rs"
 doctest = false
 
 [dependencies]
