+++ crates/assistant/README.zmd
