+++ crates/assistant/features.zmd
