+++ crates/assistant/src/assistant_panel.rs
