+++ crates/assistant/src/assistant_settings.rs
