+++ crates/assistant/src/streaming_diff.rs
